refactor!: move old trie into a separate subpackage
