+++ src/main/java/com/limechain/runtime/RuntimeBuilder.java
@@ -12,9 +12,9 @@ import com.limechain.runtime.hostapi.OffchainHostFunctions;
 import com.limechain.runtime.hostapi.StorageHostFunctions;
 import com.limechain.runtime.hostapi.TrieHostFunctions;
 import com.limechain.sync.warpsync.dto.RuntimeCodeException;
-import com.limechain.trie.Trie;
-import com.limechain.trie.TrieVerifier;
-import com.limechain.trie.decoder.TrieDecoderException;
+import com.limechain.trie.decoded.Trie;
+import com.limechain.trie.decoded.TrieVerifier;
+import com.limechain.trie.decoded.decoder.TrieDecoderException;
 import com.limechain.utils.ByteArrayUtils;
 import com.limechain.utils.LittleEndianUtils;
 import com.limechain.utils.StringUtils;
