+++ src/main/java/com/limechain/trie/decoded/NodeKind.java
@@ -1,4 +1,4 @@
-package com.limechain.trie;
+package com.limechain.trie.decoded;
 
 /**
  * Enum containing the different kind of nodes
