+++ src/main/java/com/limechain/trie/decoded/NodeVariant.java
@@ -1,4 +1,4 @@
-package com.limechain.trie;
+package com.limechain.trie.decoded;
 
 /**
  * Enum containing the different node variant bits and masks specified
