+++ src/main/java/com/limechain/trie/decoded/encoder/TrieEncoder.java
@@ -1,9 +1,9 @@
-package com.limechain.trie.encoder;
+package com.limechain.trie.decoded.encoder;
 
-import com.limechain.trie.Nibbles;
-import com.limechain.trie.Node;
-import com.limechain.trie.NodeKind;
-import com.limechain.trie.NodeVariant;
+import com.limechain.trie.decoded.Nibbles;
+import com.limechain.trie.decoded.Node;
+import com.limechain.trie.decoded.NodeKind;
+import com.limechain.trie.decoded.NodeVariant;
 import io.emeraldpay.polkaj.scale.ScaleCodecWriter;
 import io.emeraldpay.polkaj.types.Hash256;
 import lombok.experimental.UtilityClass;
@@ -11,7 +11,7 @@ import lombok.experimental.UtilityClass;
 import java.io.IOException;
 import java.io.OutputStream;
 
-import static com.limechain.trie.TrieVerifier.MAX_PARTIAL_KEY_LENGTH;
+import static com.limechain.trie.decoded.TrieVerifier.MAX_PARTIAL_KEY_LENGTH;
 
 /**
  * Encodes a {@link Node} to a {@link OutputStream} buffer.
