+++ src/test/java/com/limechain/sync/warpsync/state/RuntimeDownloadTest.java
@@ -5,9 +5,9 @@ import com.limechain.network.protocol.lightclient.LightMessages;
 import com.limechain.network.protocol.lightclient.LightMessagesProtocol;
 import com.limechain.network.protocol.lightclient.pb.LightClientMessage;
 import com.limechain.runtime.RuntimeBuilder;
-import com.limechain.trie.Trie;
-import com.limechain.trie.TrieVerifier;
-import com.limechain.trie.decoder.TrieDecoderException;
+import com.limechain.trie.decoded.Trie;
+import com.limechain.trie.decoded.TrieVerifier;
+import com.limechain.trie.decoded.decoder.TrieDecoderException;
 import com.limechain.utils.LittleEndianUtils;
 import com.limechain.utils.RandomGenerationUtils;
 import com.limechain.utils.StringUtils;
