+++ src/test/java/com/limechain/trie/decoded/Helper.java
@@ -1,6 +1,6 @@
-package com.limechain.trie;
+package com.limechain.trie.decoded;
 
-import com.limechain.trie.encoder.TrieEncoder;
+import com.limechain.trie.decoded.encoder.TrieEncoder;
 import com.limechain.utils.RandomGenerationUtils;
 
 import java.io.ByteArrayOutputStream;
